+++ SendBirdSDK.framework/Headers/SBDBaseMessageParams.h
@@ -8,6 +8,7 @@
 
 #import <Foundation/Foundation.h>
 #import "SBDTypes.h"
+#import "SBDMessageMetaArray.h"
 
 @class SBDUser;
 
@@ -37,9 +38,15 @@
 
 /**
  Meta array keys.
- @since 3.0.116
+ @deprecated in 3.0.179 Use `metaArrays` instead
  */
-@property (copy, nonatomic, nullable) NSArray<NSString *> *metaArrayKeys;
+@property (copy, nonatomic, nullable) NSArray<NSString *> *metaArrayKeys DEPRECATED_ATTRIBUTE;
+
+/**
+ Meta array with keys and values
+ @since 3.0.179
+ */
+@property (copy, nonatomic, nullable) NSArray<SBDMessageMetaArray *> *metaArrays;
 
 /**
  *